More refactoring to remove ullyses from coadd.  Had to change some class and function names, so these were also changed in timeseries.py to reflect this
