+++ ullyses/coadd.py
@@ -1,16 +1,9 @@
 import os
 import glob
-import datetime
-from datetime import datetime as dt
-import re
 
-import pandas as pd
 import numpy as np
-import astropy
 from astropy.io import fits
-from astropy.time import Time
 
-import ullyses_utils
 # coadd data
 #
 
@@ -54,7 +47,7 @@ class SegmentList:
         alldata = []
         allhdr0 = []
         allhdr1 = []
-        
+
         for file in x1dfiles:
             with fits.open(file) as f1:
                 prihdr = f1[0].header
@@ -91,7 +84,7 @@ class SegmentList:
                     alldata.append(data)
                     allhdr0.append(prihdr)
                     allhdr1.append(hdr1)
-                
+
                 if len(data) > 0:
                     print('{} added to file list for FUSE'.format(file))
                     self.instrument = 'FUSE'
@@ -237,274 +230,20 @@ class SegmentList:
         self.signal_to_noise[nonzeros] = self.output_flux[nonzeros] / self.output_errors[nonzeros]
         return
 
-    def write(self, filename, overwrite=False, level="", version=""):
-
-        # If the target is a ULLYSES target, use the official
-        # target name and coords
-        self.target = self.ull_targname()
-        self.targ_ra, self.targ_dec = self.ull_coords()
-
-        # Table 1 - HLSP data
-
-        # set up the header
-        hdr1 = fits.Header()
-        hdr1['EXTNAME'] = ('SCIENCE', 'Spectrum science arrays')
-        hdr1['TIMESYS'] = ('UTC', 'Time system in use')
-        hdr1['TIMEUNIT'] = ('s', 'Time unit for durations')
-        hdr1['TREFPOS'] = ('GEOCENTER', 'Time reference position')
-
-        mjd_beg = self.combine_keys("expstart", "min")
-        mjd_end = self.combine_keys("expend", "max")
-        dt_beg = Time(mjd_beg, format="mjd").datetime
-        dt_end = Time(mjd_end, format="mjd").datetime
-        hdr1['DATE-BEG'] = (dt.strftime(dt_beg, "%Y-%m-%dT%H:%M:%S"), 'Date-time of first observation start')
-        hdr1.add_blank('', after='TREFPOS')
-        hdr1.add_blank('              / FITS TIME COORDINATE KEYWORDS', before='DATE-BEG')
-
-        hdr1['DATE-END'] = (dt.strftime(dt_end, "%Y-%m-%dT%H:%M:%S"), 'Date-time of last observation end')
-        hdr1['MJD-BEG'] = (mjd_beg, 'MJD of first exposure start')
-        hdr1['MJD-END'] = (mjd_end, 'MJD of last exposure end')
-        hdr1['XPOSURE'] = (self.combine_keys("exptime", "sum"), '[s] Sum of exposure durations')
-
-        # set up the table columns
-        nelements = len(self.output_wavelength)
-        rpt = str(nelements)
-
-        # Table with co-added spectrum
-        cw = fits.Column(name='WAVELENGTH', format=rpt+'E', unit="Angstrom")
-        cf = fits.Column(name='FLUX', format=rpt+'E', unit="erg /s /cm**2 /Angstrom")
-        ce = fits.Column(name='ERROR', format=rpt+'E', unit="erg /s /cm**2 /Angstrom")
-        cs = fits.Column(name='SNR', format=rpt+'E')
-        ct = fits.Column(name='EFF_EXPTIME', format=rpt+'E', unit="s")
-        cd = fits.ColDefs([cw, cf, ce, cs, ct])
-        table1 = fits.BinTableHDU.from_columns(cd, nrows=1, header=hdr1)
-
-        # populate the table
-        table1.data['WAVELENGTH'] = self.output_wavelength.copy()
-        table1.data['FLUX'] = self.output_flux.copy()
-        table1.data['ERROR'] = self.output_errors.copy()
-        table1.data['SNR'] = self.signal_to_noise.copy()
-        table1.data['EFF_EXPTIME'] = self.output_exptime.copy()
-        # HLSP primary header
-        hdr0 = fits.Header()
-        hdr0['EXTEND'] = ('T', 'FITS file may contain extensions')
-        hdr0['NEXTEND'] = 3
-        hdr0['FITS_VER'] = 'Definition of the Flexible Image Transport System (FITS) v4.0 https://fits.gsfc.nasa.gov/standard40/fits_standard40aa-le.pdf'
-        hdr0['FITS_SW'] = ('astropy.io.fits v' + astropy.__version__, 'FITS file creation software')
-        hdr0['ORIGIN'] = ('Space Telescope Science Institute', 'FITS file originator')
-        hdr0['DATE'] = (str(datetime.date.today()), 'Date this file was written')
-        hdr0['FILENAME'] = (os.path.basename(filename), 'Name of this file')
-        hdr0['TELESCOP'] = (self.combine_keys("telescop", "multi"), 'Telescope used to acquire data')
-        hdr0['INSTRUME'] = (self.combine_keys("instrume", "multi"), 'Instrument used to acquire data')
-        hdr0.add_blank('', after='TELESCOP')
-        hdr0.add_blank('              / SCIENCE INSTRUMENT CONFIGURATION', before='INSTRUME')
-        hdr0['DETECTOR'] = (self.combine_keys("detector", "multi"), 'Detector or channel used to acquire data')
-        hdr0['DISPERSR'] = (self.combine_keys("opt_elem", "multi"), 'Identifier of disperser')
-        hdr0['CENWAVE'] = (self.combine_keys("cenwave", "multi"), 'Central wavelength setting for disperser')
-        hdr0['APERTURE'] = (self.combine_keys("aperture", "multi"), 'Identifier of entrance aperture')
-        hdr0['S_REGION'] = (self.obs_footprint(), 'Region footprint')
-        hdr0['OBSMODE'] = (self.combine_keys("obsmode", "multi"), 'Instrument operating mode (ACCUM | TIME-TAG)')
-        hdr0['TARGNAME'] = self.target
-        hdr0.add_blank(after='OBSMODE')
-        hdr0.add_blank('              / TARGET INFORMATION', before='TARGNAME')
-
-        hdr0['RADESYS'] = ('ICRS ','World coordinate reference frame')
-        hdr0['TARG_RA'] =  (self.targ_ra,  '[deg] Target right ascension')
-        hdr0['TARG_DEC'] =  (self.targ_dec,  '[deg] Target declination')
-        hdr0['PROPOSID'] = (self.combine_keys("proposid", "multi"), 'Program identifier')
-        hdr0.add_blank(after='TARG_DEC')
-        hdr0.add_blank('           / PROVENANCE INFORMATION', before='PROPOSID')
-        hdr0['CAL_VER'] = (f'ULLYSES Cal {CAL_VER}', 'HLSP processing software version')
-        hdr0['HLSPID'] = ('ULLYSES', 'Name ID of this HLSP collection')
-        hdr0['HSLPNAME'] = ('Hubble UV Legacy Library of Young Stars as Essential Standards',
-                        'Name ID of this HLSP collection')
-        hdr0['HLSPLEAD'] = ('Julia Roman-Duval', 'Full name of HLSP project lead')
-        hdr0['HLSP_VER'] = (version,'HLSP data release version identifier')
-        hdr0['HLSP_LVL'] = (level, 'ULLYSES HLSP Level')
-        hdr0['LICENSE'] = ('CC BY 4.0', 'License for use of these data')
-        hdr0['LICENURL'] = ('https://creativecommons.org/licenses/by/4.0/', 'Data license URL')
-        hdr0['REFERENC'] = ('https://ui.adsabs.harvard.edu/abs/2020RNAAS...4..205R', 'Bibliographic ID of primary paper')
-
-        hdr0['CENTRWV'] = (self.combine_keys("centrwv", "average"), 'Central wavelength of the data')
-        hdr0.add_blank(after='REFERENC')
-        hdr0.add_blank('           / ARCHIVE SEARCH KEYWORDS', before='CENTRWV')
-        hdr0['MINWAVE'] = (self.combine_keys("minwave", "min"), 'Minimum wavelength in spectrum')
-        hdr0['MAXWAVE'] = (self.combine_keys("maxwave", "max"), 'Maximum wavelength in spectrum')
-
-        primary = fits.PrimaryHDU(header=hdr0)
-
-        # Table 2 - individual product info
-
-        # first set up header
-        hdr2 = fits.Header()
-        hdr2['EXTNAME'] = ('PROVENANCE', 'Metadata for contributing observations')
-        # set up the table columns
-        cfn = fits.Column(name='FILENAME', array=self.combine_keys("filename", "arr"), format='A64')
-        cpid = fits.Column(name='PROPOSID', array=self.combine_keys("proposid", "arr"), format='A32')
-        ctel = fits.Column(name='TELESCOPE', array=self.combine_keys("telescop", "arr"), format='A32')
-        cins = fits.Column(name='INSTRUMENT', array=self.combine_keys("instrume", "arr"), format='A32')
-        cdet = fits.Column(name='DETECTOR', array=self.combine_keys("detector", "arr"), format='A32')
-        cdis = fits.Column(name='DISPERSER', array=self.combine_keys("opt_elem", "arr"), format='A32')
-        ccen = fits.Column(name='CENWAVE', array=self.combine_keys("cenwave", "arr"), format='A32')
-        cap = fits.Column(name='APERTURE', array=self.combine_keys("aperture", "arr"), format='A32')
-        csr = fits.Column(name='SPECRES', array=self.combine_keys("specres", "arr"), format='F8.1')
-        ccv = fits.Column(name='CAL_VER', array=self.combine_keys("cal_ver", "arr"), format='A32')
-        mjd_begs = self.combine_keys("expstart", "arr")
-        mjd_ends = self.combine_keys("expend", "arr")
-        mjd_mids = (mjd_ends + mjd_begs) / 2.
-        cdb = fits.Column(name='MJD_BEG', array=mjd_begs, format='F15.9', unit='d')
-        cdm = fits.Column(name='MJD_MID', array=mjd_mids, format='F15.9', unit='d')
-        cde = fits.Column(name='MJD_END', array=mjd_ends, format='F15.9', unit='d')
-        cexp = fits.Column(name='XPOSURE', array=self.combine_keys("exptime", "arr"), format='F15.9', unit='s')
-        cmin = fits.Column(name='MINWAVE', array=self.combine_keys("minwave", "arr"), format='F9.4', unit='Angstrom')
-        cmax = fits.Column(name='MAXWAVE', array=self.combine_keys("maxwave", "arr"), format='F9.4', unit='Angstrom')
-
-        cd2 = fits.ColDefs([cfn, cpid, ctel, cins, cdet, cdis, ccen, cap, csr, ccv, cdb, cdm, cde, cexp, cmin ,cmax])
-
-        table2 = fits.BinTableHDU.from_columns(cd2, header=hdr2)
-
-        # the HDUList:
-        # 0 - empty data - 0th ext header
-        # 1 - HLSP data - 1st ext header
-        # 2 - individual product info - 2nd ext header
-
-        hdul = fits.HDUList([primary, table1, table2])
-
-        hdul.writeto(filename, overwrite=overwrite)
-
-        # from ullyses_jira.parse_csv import parse_name_csv
-        # name_mapping = {}
-        # for ttype in ['lmc', 'smc', 'tts']:
-        #     names_dict = parse_name_csv(ttype)
-        #     name_mapping = {**name_mapping, **names_dict}
-
-
-    def obs_footprint(self):
-        # Not using WCS at the moment
-        # This is a placeholder, need to figure out polygon
-#        apertures = list(set([h["aperture"] for h in self.primary_headers]))
-#        ras = list(set([h["ra_targ"] for h in self.primary_headers]))
-#        ra_diff = max(np.abs(ras)) - min(np.abs(ras))
-#        decs = list(set([h["dec_targ"] for h in self.primary_headers]))
-#        dec_diff = max(np.abs(decs)) - min(np.abs(decs))
-#        center_ra = np.average(ras)
-#        center_dec = np.average(decs)
-#        extent_ra = (2.5 / 2 / 3600) + ra_diff
-#        extent_dec = (2.5 / 2 / 3600) + dec_diff
-#        radius = max([extent_ra, extent_dec])
-        self.targ_ra, self.targ_dec = self.ull_coords()
-        radius = (2.5 / 2 / 3600)
-        center_ra = self.targ_ra
-        center_dec = self.targ_dec
-
-        s_region = f"CIRCLE {center_ra} {center_dec} {radius}"
-        return s_region
-
-    def ull_targname(self):
-        aliases_file = ullyses_utils.__path__[0] + '/data/target_metadata/pd_all_aliases.json'
-        aliases = pd.read_json(aliases_file, orient="split")
+    def get_targname(self):
         # These are just preliminary target names, in case we can't find a match
         if len(self.targnames) == 1:
-            ull_targname = self.targnames[0]
+            targname = self.targnames[0]
         else:
-            ull_targname = self.primary_headers[0]["targname"]
-        targ_matched = False
-        for targ in self.targnames:
-            # The alias file is all uppercase
-            targ_upper = targ.upper()
-            mask = aliases.apply(lambda row: row.astype(str).str.fullmatch(re.escape(targ_upper)).any(), axis=1)
-            if set(mask) != {False}:
-                targ_matched = True
-                ull_targname = aliases[mask]["ULL_MAST_name"].values[0]
-                break
-        if targ_matched is False:
-            print(f"{RED}WARNING: Could not match target name {ull_targname} to ULLYSES alias list{RESET}")
-        return ull_targname
+            targname = self.primary_headers[0]["targname"]
+        return targname
 
-    def ull_coords(self):
+    def get_coords(self):
         ras = list(set([h["ra_targ"] for h in self.primary_headers]))
         decs = list(set([h["dec_targ"] for h in self.primary_headers]))
         avg_ra = np.average(ras)
         avg_dec = np.average(decs)
-        if self.target == "":
-            return avg_ra, avg_dec
-        targetinfo_file = ullyses_utils.__path__[0] + '/data/target_metadata/pd_targetinfo.json'
-        master_list = pd.read_json(targetinfo_file, orient="split")
-        master_list = master_list.apply(lambda x: x.astype(str).str.upper())
-        coords = master_list.loc[master_list["mast_targname"] == self.target][["ra", "dec"]].values
-        if len(coords) != 0:
-            return coords[0][0], coords[0][1]
-        else:
-            return avg_ra, avg_dec
-
-
-    def combine_keys(self, key, method):
-        keymap= {"HST": {"expstart": ("expstart", 1),
-                         "expend": ("expend", 1),
-                         "exptime": ("exptime", 1),
-                         "telescop": ("telescop", 0),
-                         "instrume": ("instrume", 0),
-                         "detector": ("detector", 0),
-                         "opt_elem": ("opt_elem", 0),
-                         "cenwave": ("cenwave", 0),
-                         "aperture": ("aperture", 0),
-                         "obsmode": ("obsmode", 0),
-                         "proposid": ("proposid", 0),
-                         "centrwv": ("centrwv", 0),
-                         "minwave": ("minwave", 0),
-                         "maxwave": ("maxwave", 0),
-                         "filename": ("filename", 0),
-                         "specres": ("specres", 0),
-                         "cal_ver": ("cal_ver", 0)},
-                "FUSE": {"expstart": ("obsstart", 0),
-                         "expend": ("obsend", 0),
-                         "exptime": ("obstime", 0),
-                         "telescop": ("telescop", 0),
-                         "instrume": ("instrume", 0),
-                         "detector": ("detector", 0),
-                         "opt_elem": ("detector", 0),
-                         "cenwave": ("centrwv", 0),
-                         "aperture": ("aperture", 0),
-                         "obsmode": ("instmode", 0),
-                         "proposid": ("prgrm_id", 0),
-                         "centrwv": ("centrwv", 0),
-                         "minwave": ("wavemin", 0),
-                         "maxwave": ("wavemax", 0),
-                         "filename": ("filename", 0),
-                         "specres": ("spec_rp", 1),
-                         "cal_ver": ("cf_vers", 0)}}
-
-        vals = []
-        for i in range(len(self.primary_headers)):
-            tel = self.primary_headers[i]["telescop"]
-            actual_key = keymap[tel][key][0]
-            hdrno = keymap[tel][key][1]
-            if hdrno == 0:
-                val = self.primary_headers[i][actual_key]
-            else:
-                val = self.first_headers[i][actual_key]
-            if tel == "FUSE" and key == "filename":
-                val = val.replace(".fit", "_vo.fits")
-            vals.append(val)
-
-        # Allowable methods are min, max, average, sum, multi, arr
-        if method == "multi":
-            keys_set = list(set(vals))
-            if len(keys_set) > 1:
-                return "MULTI"
-            else:
-                return keys_set[0]
-        elif method == "min":
-            return min(vals)
-        elif method == "max":
-            return max(vals)
-        elif method == "average":
-            return np.average(vals)
-        elif method == "sum":
-            return np.sum(vals)
-        elif method == "arr":
-            return np.array(vals)
-
+        return avg_ra, avg_dec
 
 # Weight functions for STIS
 weight_function = {
@@ -747,7 +486,7 @@ def abut(product_short, product_long):
             long_indices = np.where(product_long.output_wavelength > first_good_long)
             transition_index_long = long_indices[0][0]
         output_grating = product_short.grating + '-' + product_long.grating
-        product_abutted = SegmentList('', output_grating)
+        product_abutted = product_short.__class__('', output_grating)
         nout = len(product_short.output_wavelength[:transition_index_short])
         nout = nout + len(product_long.output_wavelength[transition_index_long:])
         product_abutted.nelements = nout
@@ -769,8 +508,8 @@ def abut(product_short, product_long):
         product_abutted.primary_headers = product_short.primary_headers + product_long.primary_headers
         product_abutted.first_headers = product_short.first_headers + product_long.first_headers
         product_abutted.grating = output_grating
-        product_short.target = product_short.ull_targname()
-        product_long.target = product_long.ull_targname()
+        product_short.target = product_short.get_targname()
+        product_long.target = product_long.get_targname()
         if product_short.instrument in product_long.instrument:
             product_abutted.instrument = product_long.instrument
         if product_long.instrument in product_short.instrument:
@@ -785,7 +524,7 @@ def abut(product_short, product_long):
         else:
             for target_name in product_short.targnames:
                 if target_name in product_long.targnames:
-                    product_abutted.target = product_abutted.ull_targname()
+                    product_abutted.target = product_abutted.get_targname()
                     target_matched = True
                     product_abutted.targnames = [target_name]
         if not target_matched:
@@ -799,7 +538,7 @@ def abut(product_short, product_long):
             product_abutted = product_long
         else:
             product_abutted = None
-    product_abutted.targ_ra, product_abutted.targ_dec = product_abutted.ull_coords()
+    product_abutted.targ_ra, product_abutted.targ_dec = product_abutted.get_coords()
     return product_abutted
